[TLC-380] Renaming, changes as per review feedback
