+++ src/app/item-page/edit-item-page/edit-item-page.routing.module.ts
@@ -10,7 +10,7 @@ import { ItemStatusComponent } from './item-status/item-status.component';
 import { ItemBitstreamsComponent } from './item-bitstreams/item-bitstreams.component';
 import { ItemCollectionMapperComponent } from './item-collection-mapper/item-collection-mapper.component';
 import { ItemMoveComponent } from './item-move/item-move.component';
-import { ItemRegisterDoiComponent } from './item-register-doi/item-registerdoi.component';
+import { ItemRegisterDoiComponent } from './item-register-doi/item-register-doi.component';
 import { ItemRelationshipsComponent } from './item-relationships/item-relationships.component';
 import { I18nBreadcrumbResolver } from '../../core/breadcrumbs/i18n-breadcrumb.resolver';
 import { ItemVersionHistoryComponent } from './item-version-history/item-version-history.component';
@@ -40,7 +40,7 @@ import { ItemPageRelationshipsGuard } from './item-page-relationships.guard';
 import { ItemPageVersionHistoryGuard } from './item-page-version-history.guard';
 import { ItemPageCollectionMapperGuard } from './item-page-collection-mapper.guard';
 import { ThemedDsoEditMetadataComponent } from '../../dso-shared/dso-edit-metadata/themed-dso-edit-metadata.component';
-import { ItemPageRegisterDoiGuard } from './item-page-registerdoi.guard';
+import { ItemPageRegisterDoiGuard } from './item-page-register-doi.guard';
 
 /**
  * Routing module that handles the routing for the Edit Item page administrator functionality
@@ -149,7 +149,7 @@ import { ItemPageRegisterDoiGuard } from './item-page-registerdoi.guard';
             path: ITEM_EDIT_REGISTER_DOI_PATH,
             component: ItemRegisterDoiComponent,
             canActivate: [ItemPageRegisterDoiGuard],
-            data: { title: 'item.edit.registerdoi.title' },
+            data: { title: 'item.edit.register-doi.title' },
           },
           {
             path: ITEM_EDIT_AUTHORIZATIONS_PATH,
