+++ src/app/item-page/edit-item-page/item-register-doi/item-register-doi-component.html
